requirements.txt
streamlit>=1.22.0
newsapi-python>=0.2.6
python-dotenv>=0.21.0
google-generativeai>=0.3.0
requests>=2.28.1
pandas>=1.4.4
pillow>=9.2.0 